snake movement packets

+++ src/game/world.hpp
@@ -32,7 +32,7 @@ public:
     void flush_changes();
 
     // world
-    const uint32_t game_radius = 21600;
+    const uint16_t game_radius = 21600;
     const uint16_t max_snake_parts = 411;
     const uint16_t sector_size = 300;
     const uint16_t sector_count_along_edge = 144;
